Doesn't return relationships that don't exist anymore in DB or DS.Store
There are cases where a given record was deleted from the DB, but its
parent still has a reference to its ID. In that case, Ember.js will get
the reference ID, but the record doesn't really exist, and it will blow.

In this commit, we iterate over the relationships and check if it exists
either in the DS.Store or in the embedded records (inside IndexedDB). If
it doesn't exist anywhere, we simply not return the ID of the missing
record.